Module 1
added meta element naming myself as the author of the page.
changed the title to display a name and job title.
wrapped the existing text in paragraph tags.
added a new paragraph with a brief message to your users.

added a "useful links used" section
added a "for fun" section
added a picture